requirements.txt
certifi==2024.6.2
cycler==0.12.1
future==1.0.0
kiwisolver==1.4.5
matplotlib==3.3.1
numpy==1.19.1
opencv-contrib-python==4.4.0.44
Pillow==8.0.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.49.0
